requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Redis for timer and buzzer
redis==5.0.1
aioredis==2.0.1

# WebSocket
websockets==12.0

# PPT Processing
python-pptx==0.6.23
Pillow==10.2.0
pdf2image==1.17.0

# QR Code generation
qrcode[pil]==7.4.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
